ai_docs_backend/requirements.txt
fastapi
uvicorn
openai
beautifulsoup4
requests
python-dotenv
pdfplumber
chromadb
tiktoken